+++ shaders/2d_projection.vert
@@ -1,6 +1,6 @@
 #version 430 core
 
-layout (location = 0) in ivec2 in_position;
+layout (location = 0) in vec2 in_position;
 layout (location = 1) in vec4 in_color;
 
 out vec4 vert_color;
